refactor(ui): move fully-duplicated RunDetailsV3/V4 siblings into RunDetails/shared
IO, Nav, ErrorInfo, Waiting, CancelModal were byte-identical between the
two folders; Tabs.tsx's V4 variant was a clean superset (wider label type,
safer equality, additional flex classes). Relocated all six into
ui/packages/components/src/RunDetails/shared/ and rewrote the ~12
importers. No behavioral change.
