+++ ui/packages/components/src/RunDetails/shared/ErrorInfo.tsx
