+++ ui/packages/components/src/RunDetailsV4/StepInfo.tsx
@@ -22,10 +22,10 @@ import { Time } from '../Time';
 import type { SpanMetadataKind } from '../generated';
 import { usePrettyErrorBody, usePrettyJson, usePrettyShortError } from '../hooks/usePrettyJson';
 import { toMaybeDate } from '../utils/date';
-import { ErrorInfo } from './ErrorInfo';
-import { IO } from './IO';
+import { ErrorInfo } from '../RunDetails/shared/ErrorInfo';
+import { IO } from '../RunDetails/shared/IO';
 import { MetadataAttrs } from './MetadataAttrs';
-import { Tabs } from './Tabs';
+import { Tabs } from '../RunDetails/shared/Tabs';
 import { UserlandAttrs } from './UserlandAttrs';
 import { formatDuration, maybeBooleanToString, type StepInfoType } from './runDetailsUtils';
 import {
